refactor: rename "template-reports" to "manager" component :hammer:
