+++ components/manager/.env.example
