+++ components/manager/Dockerfile
