+++ components/manager/Makefile
