+++ components/manager/api/swagger.json
