+++ components/manager/api/swagger.yaml
