+++ components/manager/docker-compose.yml
