+++ components/manager/internal/adapters/http/in/template_handler.go
@@ -3,7 +3,7 @@ package in
 import (
 	"github.com/flosch/pongo2/v6"
 	"github.com/gofiber/fiber/v2"
-	"plugin-template-engine/components/template-reports/internal/services"
+	"plugin-template-engine/components/manager/internal/services"
 	"plugin-template-engine/pkg"
 	"plugin-template-engine/pkg/net/http"
 	"plugin-template-engine/pkg/opentelemetry"
