+++ components/manager/internal/adapters/http/in/template_middlewares.go
