+++ components/manager/internal/adapters/mongodb/templates/template.go
