+++ components/manager/internal/adapters/mongodb/templates/template.mongodb.go
