--- components/template-reports/grafana/run-grafana.sh
@@ -1,12 +1,0 @@
-#!/bin/bash
-
-source ./logging.sh
-
-export GF_AUTH_ANONYMOUS_ENABLED=false
-
-export GF_PATHS_HOME=/data/grafana
-export GF_PATHS_DATA=/data/grafana/data
-export GF_PATHS_PLUGINS=/data/grafana/plugins
-
-cd ./grafana || exit
-run_with_logging "Grafana ${GRAFANA_VERSION}" "${ENABLE_LOGS_GRAFANA:-false}" ./bin/grafana server
